CI: tweaks (#113)
* CI: tweaks

* explicitly choose CXX=g++-10

* updates for upcoming pycvodes

* _config.env => _config

* use dev option: pip install -e

* pip opts

* pytest: norecursedirs += tmp

* get_link_libs() -> get_libs()

* bump dependencies (old versions are no longer tested)

* sundials install does not contain pkg-info nor cmake config, we need to set CPATH

* LD_LIBRARY_PATH...

* updates

* forgot one env -> config conversion

* KLU variable fix

* handle __getitem__ in MissingImport

* update merge blocker script

* Revert "update merge blocker script"

This reverts commit 2443b4a3033ba1f1c837174003e5bc93ea2cde3e.

* reorder

* py2 compat.

* py2: mpl

* g++-10

* skip py2 tests for now

* use pycvodes-0.13.0

+++ .ci/run_ci.sh
@@ -1,18 +1,31 @@
 #!/bin/bash -xe
 
+set -u
 export PKG_NAME=$1
+SUNDBASE=$2
+set +u
 
-for p in "${@:2}"
-do
-export CPATH=$p/include:$CPATH LIBRARY_PATH=$p/lib:$LIBRARY_PATH LD_LIBRARY_PATH=$p/lib:$LD_LIBRARY_PATH
-done
+if [ ! -e "$SUNDBASE/include/sundials/sundials_config.h" ]; then
+    >&2 echo "Not a valid prefix for sundials: $SUNDBASE"
+    exit 1
+fi
 
+mkdir -p $HOME/.config/pip/
+echo -e "[global]\nno-cache-dir = false\ndownload-cache = $(pwd)/ci_cache/pip_cache" >$HOME/.config/pip/pip.conf
 python3 -m pip install symcxx pysym  # unofficial backends, symengine is tested in the conda build
+
+# (cd ./tmp/pycvodes;
+CFLAGS="-isystem $SUNDBASE/include $CFLAGS" LDFLAGS="-Wl,--disable-new-dtags -Wl,-rpath,$SUNDBASE/lib -L$SUNDBASE/lib $LDFLAGS" python3 -m pip install pycvodes # setup.py install )
+git clean -xfd # -e tmp/
+
+export CPATH=$SUNDBASE/include
+export LIBRARY_PATH=$SUNDBASE/lib
+export LD_LIBRARY_PATH=$SUNDBASE/lib
 
 python3 setup.py sdist
 PKG_VERSION=$(python3 setup.py --version)
 (cd dist/; python3 -m pip install $PKG_NAME-$PKG_VERSION.tar.gz)
-python3 -m pip install --upgrade --upgrade-strategy only-if-needed .[all]
+python3 -m pip install -e .[all]
 python3 -m pytest -xv -k test_integrate_chained_robertson pyodesys/tests/test_robertson.py
 export PYTHONHASHSEED=$(python3 -c "import random; print(random.randint(1,2**32-1))")
 PYTHON="python3 -R" ./scripts/run_tests.sh --cov $PKG_NAME --cov-report html
